+++ src.mk
@@ -95,6 +95,7 @@ LIB_SOURCES =                                                   \
   util/env.cc                                                   \
   util/env_hdfs.cc                                              \
   util/env_posix.cc                                             \
+  util/io_posix.cc                                              \
   util/file_util.cc                                             \
   util/file_reader_writer.cc                                    \
   util/filter_policy.cc                                         \
